Исправлено формирования SEO паттернов компонентов при включенной мультиязычности; Для TinyMCE добавлены опции "Блочный элемент" (вместо Перевод каретки) и "Поведение новой строки"; Для комментариев добавлена опция "Скрывать комментарии удалённых пользователей"; Добавлена поддержка мультиязычности для связки Запись ТК -> Группа

+++ system/controllers/comments/hooks/user_delete.php
@@ -2,12 +2,13 @@
 
 class onCommentsUserDelete extends cmsAction {
 
-    public function run($user){
+    public function run($user) {
 
-        $this->model->deleteUserComments($user['id']);
+        if (!empty($this->options['hide_deleted_user_comments'])) {
+            $this->model->deleteUserComments($user['id']);
+        }
 
         return $user;
-
     }
 
 }
